+++ Frends.Files.LocalBackup/Frends.Files.LocalBackup/Definitions/Input.cs
@@ -1,4 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel;
 
 namespace Frends.Files.LocalBackup.Definitions
 {
@@ -22,11 +23,18 @@ namespace Frends.Files.LocalBackup.Definitions
         public string SourceDirectory { get; set; }
 
         /// <summary>
-        /// Destination directory where backup folder will be created. Backup directory's format: {BackupDirectory}{timestamp}-{Guid}.
+        /// Destination directory where backup folder will be created. Backup directory's format: {BackupDirectory}-{Guid}.
         /// </summary>
         /// <example>c:\temp\backups</example>
         [DisplayFormat(DataFormatString = "Text")]
         public string BackupDirectory { get; set; }
+
+        /// <summary>
+        /// Execution id of the task which is used in naming the backup directory.
+        /// </summary>
+        /// <example>e7e34166-f4fd-45e5-9307-ea5c2cf8e037</example>
+        [DefaultValue("#process.executionid")]
+        public string TaskExecutionId { get; set; }
 
         /// <summary>
         /// Cleanup older than {DaysOlder} folders from backup directory.