+++ Frends.Files.LocalBackup/Frends.Files.LocalBackup/LocalBackup.cs
@@ -28,22 +28,27 @@ namespace Frends.Files.LocalBackup
             if (string.IsNullOrWhiteSpace(input.SourceDirectory) || string.IsNullOrWhiteSpace(input.SourceFile)) throw new Exception("Source parameters required.");
             if (string.IsNullOrWhiteSpace(input.BackupDirectory)) throw new Exception("Backup directory required.");
 
-            var backup = CreateBackup(input, cancellationToken);
+            var (directory, backup) = CreateBackup(input, cancellationToken);
+
             var cleanup = input.Cleanup ? CleanUp(input, cancellationToken) : null;
 
-            return new Result(backup, cleanup);
+            return new Result(directory, backup.Count, backup, cleanup);
         }
 
-        private static List<BackupObject> CreateBackup(Input input, CancellationToken cancellationToken)
+        private static Tuple<string, List<BackupObject>> CreateBackup(Input input, CancellationToken cancellationToken)
         {
             var timestampString = DateTime.UtcNow.ToString("yyyy-MM-dd_HH_mm_ss");
+
             var result = new List<BackupObject>();
 
+            var guid = (!string.IsNullOrEmpty(input.TaskExecutionId)) ? input.TaskExecutionId : Guid.NewGuid().ToString();
+
             //Create backup folder
-            var backupDirectory = Path.Combine(input.BackupDirectory, $"{timestampString}-{Guid.NewGuid()}");
+            var backupDirectory = Path.Combine(input.BackupDirectory, $"{timestampString}{guid}");
             Directory.CreateDirectory(backupDirectory);
 
             var files = Directory.GetFiles(input.SourceDirectory);
+
             foreach (string file in files)
             {
                 cancellationToken.ThrowIfCancellationRequested();
@@ -56,7 +61,13 @@ namespace Frends.Files.LocalBackup
                 }
             }
 
-            return result;
+            if (!Directory.GetFiles(backupDirectory).Any())
+            {
+                Directory.Delete(backupDirectory, true);
+                return new Tuple<string, List<BackupObject>>("No source files present to backup.", new List<BackupObject>());
+            }
+
+            return new Tuple<string, List<BackupObject>>(backupDirectory, result);
         }
 
         private static List<BackupObject> CleanUp(Input input, CancellationToken cancellationToken)
@@ -68,7 +79,7 @@ namespace Frends.Files.LocalBackup
             if (input.DaysOlder < 0)
                 throw new ArgumentException("Days older cannot be a negative number", "daysOlder");
 
-            var directories = Directory.GetDirectories(input.BackupDirectory).Where(dirName => IsOlderThan(dirName, input.DaysOlder));
+            var directories = Directory.GetDirectories(input.BackupDirectory).Where(dirName => IsOlderThan(dirName, input));
             foreach (var dir in directories)
             {
                 cancellationToken.ThrowIfCancellationRequested();
@@ -78,10 +89,10 @@ namespace Frends.Files.LocalBackup
             return result;
         }
 
-        private static bool IsOlderThan(string dirPath, int daysOlder)
+        private static bool IsOlderThan(string dirPath, Input input)
         {
             var TimestampPattern = "yyyy-MM-dd_HH_mm_ss";
-            var DirNamePatternLength = TimestampPattern.Length + 37; //TimestampPattern.Length+GUID
+            var DirNamePatternLength = TimestampPattern.Length + input.TaskExecutionId.Length; //TimestampPattern.Length+GUID
             var dirName = Path.GetFileName(dirPath) ?? string.Empty;
 
             if (dirName.Length == DirNamePatternLength)
@@ -89,10 +100,12 @@ namespace Frends.Files.LocalBackup
                 var timestampPart = dirName.Substring(0, TimestampPattern.Length);
 
                 if (DateTime.TryParseExact(timestampPart, TimestampPattern, CultureInfo.InvariantCulture, DateTimeStyles.None, out DateTime timeStamp))
-                    return timeStamp < DateTime.UtcNow.AddDays(-daysOlder);
+                {
+                    return timeStamp < DateTime.UtcNow.AddDays(-input.DaysOlder);
+                }
             }
+            return new FileInfo(dirPath).LastWriteTime < DateTime.UtcNow.AddDays(-input.DaysOlder);
 
-            return new FileInfo(dirPath).LastWriteTime < DateTime.Now.AddDays(-daysOlder);
         }
 
         private static bool FileMatchesMask(string filename, string mask)