+++ Frends.Files.LocalBackup/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.0.1] - 2022-07-07
+### Added
+- Modified the task to use #process.executionid instead of a random guid
+- Added input for TaskExecutionId as guid
+- Added backup directory value to Result object
+- Added file count how many were made backup of
+
 ## [1.0.0] - 2022-05-16
 ### Added
 - Initial implementation